% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvg_gadget.R
\name{rvg_addin}
\alias{rvg_addin}
\title{Export current plot to Word or PowerPoint document.}
\usage{
rvg_addin()
}
\description{
Create a Microsoft Word or PowerPoint with the current plot
exported in a editable vector graphics format.
}
\examples{
if (interactive()){
  plot(rnorm(100))
  rvg_addin()
}
}

